services/trades/pyproject.toml
[project]
name = "trades"
version = "0.1.0"
description = "Trades microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "pydantic>=2.9.2",
    "quixstreams>=3.3.0",
    "websocket-client>=1.8.0",
]
